Automatically fix merged toggle mods
Story: For merged mods, previously users would have to go
run the program in each individual mod, then make a custom
mapping function in the merged.ini to each individual
generated `RaidenShogunRemapBlend.buf` files

Now the user can just run the program where the merged.ini is located.

This change is also compatible with mods that are not merged.

Note: Although it would be better practice to put the individual
classes into seperate files, we stick all our classes into
one file to keep the simple UI of the user copying the program into
their desired mod and running the program. It would be more organized
to seperate out the classes, but another UI may need to be introduced.